Update README.md
The planning, hiring, training, controlling, and coordinating of a company's sales team is known as sales management. Whether you are selling a service or a product, sales managers are in charge of overseeing the sales force, establishing team goals, and ultimately ensuring that the team's vision is carried out. Since it directly affects a business's revenue, a sales team's function within an organisation is possibly the most important.

Sales management software is a centralised system that aids in the implementation of sales strategies, management of sales operations, and increasing sales.

Taking charge of a sales staff is no easy task. You can find yourself getting entangled in a maze of different tools as you look for good sales management software to help you track KPIs, assign leads to sales teams, monitor the sales pipeline, and analyse lost transactions. Finally, the main goal of having quick access to sales information is compromised, and productivity plummets.

A good CRM is a comprehensive piece of software that addresses the problems of sales management procedures. Although not every CRM programme is useful for a sales manager, owning one like OneHash sales CRM for your company can significantly improve sales practices and performance by allowing sales managers to maximise their sales management talents.